separate PersesThemeProvider to support apps that do not use MUI
