+++ ui/components/src/StatChart/StatChart.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { ChartsThemeProvider } from '../context/ChartsThemeProvider';
+import { PersesThemeProvider } from '../context/PersesThemeProvider';
 import { UnitOptions } from '../model';
 import { StatChart, StatChartData } from './StatChart';
 
@@ -33,14 +33,14 @@ describe('StatChart', () => {
   describe('Render default options (no sparkline)', () => {
     it('should render', () => {
       render(
-        <ChartsThemeProvider themeName="perses">
+        <PersesThemeProvider themeName="perses">
           <StatChart
             width={contentDimensions.width}
             height={contentDimensions.height}
             data={mockStatData}
             unit={unit}
           />
-        </ChartsThemeProvider>
+        </PersesThemeProvider>
       );
       expect(screen.getByText('7.73')).toBeInTheDocument();
     });
