+++ ui/components/src/index.ts
@@ -16,8 +16,8 @@ export * from './ErrorAlert';
 export * from './ErrorBoundary';
 export * from './GaugeChart';
 export * from './LineChart';
-// export { StatChart } from './StatChart/StatChart';
 export * from './StatChart/StatChart';
 export * from './context/ChartsThemeProvider';
+export * from './context/PersesThemeProvider';
 export * from './utils/combine-sx';
 export * from './model';
