+++ ui/components/src/model/theme.ts
@@ -11,9 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ThemeOptions as MaterialThemeOptions } from '@mui/material';
 import type { EChartsOption, EChartsCoreOption, BarSeriesOption, LineSeriesOption } from 'echarts';
-import merge from 'lodash/merge';
 
 export interface PersesChartsTheme {
   themeName: string;
@@ -25,160 +23,4 @@ export interface PersesChartsTheme {
 export interface EChartsTheme extends EChartsOption {
   bar?: BarSeriesOption;
   line?: LineSeriesOption;
-}
-
-export function generateChartsTheme(
-  themeName: string,
-  muiTheme?: MaterialThemeOptions,
-  echartsTheme?: EChartsTheme
-): PersesChartsTheme {
-  const ltGrey = muiTheme?.palette?.grey ? muiTheme.palette.grey[300] : '#dee2e6';
-  const mdGrey = muiTheme?.palette?.grey ? muiTheme.palette.grey[600] : '#545454';
-  const primaryTextColor = muiTheme?.palette?.text?.primary ?? '#222';
-
-  const defaultChartsTheme = {
-    title: {
-      show: false,
-    },
-    textStyle: {
-      color: primaryTextColor,
-      fontFamily: '"Lato", sans-serif',
-      fontSize: 14,
-      // fontSize: muiTheme?.typography?.fontSize ?? 14, // TODO (sjcobb): why type incorrect from theme common.ts
-      // lineHeight: 18,
-      // fontWeight: 600,
-    },
-    grid: {
-      top: 5,
-      right: 20,
-      bottom: 0,
-      left: 20,
-      containLabel: true,
-    },
-    color: ['#8dd3c7', '#bebada', '#fb8072', '#80b1d3', '#fdb462'],
-    categoryAxis: {
-      show: true,
-      axisLabel: {
-        show: true,
-        color: primaryTextColor,
-        margin: 15,
-      },
-      axisTick: {
-        show: false,
-        length: 6,
-        lineStyle: {
-          color: mdGrey,
-        },
-      },
-      axisLine: {
-        show: true,
-        lineStyle: {
-          color: mdGrey,
-        },
-      },
-      splitLine: {
-        show: true,
-        lineStyle: {
-          color: [ltGrey],
-        },
-      },
-      splitArea: {
-        show: false,
-        areaStyle: {
-          color: [ltGrey],
-        },
-      },
-    },
-    valueAxis: {
-      show: true,
-      axisLabel: {
-        color: primaryTextColor,
-        margin: 12,
-      },
-      axisLine: {
-        show: false,
-      },
-      splitLine: {
-        show: true,
-        lineStyle: {
-          width: 0.5,
-          color: ltGrey,
-          opacity: 0.9,
-        },
-      },
-    },
-    legend: {
-      show: false,
-      type: 'scroll',
-      bottom: 0,
-      textStyle: {
-        color: primaryTextColor,
-      },
-      pageTextStyle: {
-        color: mdGrey,
-      },
-      pageIconColor: muiTheme?.palette?.action?.active,
-      pageIconInactiveColor: muiTheme?.palette?.action?.disabled,
-    },
-    toolbox: {
-      show: true,
-      top: 10,
-      right: 10,
-      iconStyle: {
-        borderColor: primaryTextColor,
-      },
-      emphasis: {
-        iconStyle: {
-          textFill: primaryTextColor,
-        },
-      },
-    },
-    tooltip: {},
-    line: {
-      showSymbol: false,
-      symbol: 'circle',
-      symbolSize: 4,
-      smooth: false,
-      lineStyle: {
-        width: 1,
-      },
-      emphasis: {
-        lineStyle: {
-          width: 1.5,
-        },
-      },
-    },
-    bar: {
-      barMaxWidth: 150,
-      itemStyle: {
-        barBorderWidth: 0,
-        barBorderColor: ltGrey,
-      },
-    },
-  };
-
-  return {
-    themeName,
-    theme: merge(defaultChartsTheme, echartsTheme),
-    noDataOption: {
-      title: {
-        show: true,
-        textStyle: {
-          color: primaryTextColor,
-          fontSize: 16,
-          fontWeight: 400,
-        },
-        text: 'No data',
-        left: 'center',
-        top: 'center',
-      },
-      xAxis: {
-        show: false,
-      },
-      yAxis: {
-        show: false,
-      },
-      // series: [],
-    },
-  };
 }
